A4: move diagnostic embeds + helpers into services/
Move cogs/diagnostic/_platform_embeds.py (2,280 lines) ->
services/diagnostic_embeds.py and cogs/diagnostic/_helpers.py ->
services/diagnostic_helpers.py. Repoint the views/diagnostic/ importers
(hub_panel, platform_panel, paginator) to the new services/ modules,
clearing the two known views -> cogs layer-boundary warnings.

The old cog module paths remain as thin re-export shims so every
existing importer and unittest.mock.patch('cogs.diagnostic._platform_embeds....')
test patch site continues to resolve. Behaviour is identical.

Fleet run -- docs/planning/ultracode-fleet-plan-2026-06-19.md
